+++ script/conversion_employer_renewal_notice_generator.rb
@@ -1,9 +1,7 @@
 require 'csv'
 
 def create_directory(path)
-  if Dir.exist?(path)
-    FileUtils.rm_rf(path)
-  end
+  FileUtils.rm_rf(path) if Dir.exist?(path)
   Dir.mkdir path
 end
 
@@ -15,7 +13,5 @@ CSV.foreach("ConversionMailing7_28_16.csv", headers: :true) do |row|
   notice_builder = ShopEmployerNotices::ConversionEmployerRenewalNotice.new(data_row)
   notice_builder.deliver
   count += 1
-  if count > 10
-    break
-  end
+  break if count > 10
 end
